Decouple RateProviderFactory with RateFetcher
